requirements.txt
Django==4.1.5
psycopg2-binary
gunicorn
python-dotenv
python-telegram-bot==20.3
requests